chore: Create LICENSE
The upstream author @sunnyark has granted this project the license "Feel free to use the source without any restrictions." I believe this includes permission for re-licensing.

Based on my personal preference (and for consistency with the a1111/sd-WebUI project), I have decided to change the license to AGPLv3.

This will not have any impact on end users. However, if you copy some code from this project elsewhere or fork and modify it, please pay attention to the licensing terms.

I believe my contributions to this project are substantial enough to justify relicensing.

Signed-off-by: Jim Chen <Jim@ChenJ.im>